test(iroh): move tests into the new test/{unit,integration} layout
Match master's reorganized test tree: the pure ticket/config tests go under
test/unit/, the in-process tunnel handshake test under test/integration/. Import
paths updated to ../../src. They're picked up by the main `test` glob and the
categorized test:unit / test:integration scripts.
